algorithmic-trading-system/requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
tensorflow>=2.13.0
keras>=2.13.0
scikit-optimize>=0.9.0

# Financial Data
yfinance>=0.2.20
alpaca-trade-api>=3.0.0
pandas-ta>=0.3.14b

# Data Processing
ta-lib>=0.4.25
pyfolio>=0.9.2

# Web Framework
streamlit>=1.25.0
plotly>=5.15.0
dash>=2.13.0

# Async & Parallel Processing
asyncio
aiohttp>=3.8.0
concurrent-futures

# Database
sqlite3

# Environment & Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Logging & Monitoring
logging
structlog>=23.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: Advanced Analytics
scipy>=1.11.0
statsmodels>=0.14.0
arch>=5.5.0